fix small business logic problems

+++ Models/Database/GeoJsonPoint.cs
@@ -8,13 +8,13 @@ public class GeoJsonPoint
     public string Type { get; set; } = "Point";
 
     // [longitude, latitude]
-    public double[] Coordinates { get; set; } = [];
+    public decimal[] Coordinates { get; set; } = [];
     
     [BsonIgnore]
     [JsonIgnore]
-    public double Lat => Coordinates[1];
+    public decimal Lat => Coordinates[1];
     
     [BsonIgnore]
     [JsonIgnore]
-    public double Lng => Coordinates[0];
+    public decimal Lng => Coordinates[0];
 }
